Store recent info in background context
Solves updating recent info when popup gets prematurely closed,
e.g. by the "foreign origin" confirmation dialog
